Load plugin variables
This commit enables a tmux config to load plugin variables to be used in
tmux-nova segments.
This works by loading a file before plugin declarations that sets up a
status-left option with the variables we want to load, then after the
plugins are sourced we load those values into tmux environment
variables to be used in segments.